Move mock-xhr helper out of test directory
Jest was picking up mock-xhr.ts as a test suite and failing because it
contains no tests. Moving it to test-utils/ avoids the test file
pattern match.
