+++ packages/media-utils/src/utils/test/sideload-to-server.ts
@@ -2,7 +2,11 @@
  * Internal dependencies
  */
 import { sideloadToServer } from '../sideload-to-server';
-import { xhrState, installMockXhr, uninstallMockXhr } from './mock-xhr';
+import {
+	xhrState,
+	installMockXhr,
+	uninstallMockXhr,
+} from '../test-utils/mock-xhr';
 
 // Mock apiFetch as both a callable function and an object with nonceMiddleware.
 const mockApiFetch = jest.fn() as jest.Mock & {
